requirements/prod.txt
gunicorn
Flask
Flask-Caching
Flask-Migrate
Flask-Script
Flask-Moment
Flask-SQLAlchemy
flask-restx
flask-marshmallow
flask-jwt-oidc
python-dotenv
psycopg2-binary
marshmallow-sqlalchemy
jsonschema
requests
croniter
attrs
jaeger-client
dpath
Werkzeug
cryptography
sqlalchemy_utils
sqlalchemy<1.4
